financial-assistant/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
